super_resolution/super_resolution_requirements.txt
# Super Resolution Standalone Module - Dependencies

torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
opencv-python>=4.5.0
Pillow>=8.0.0

# Optional: For better performance
# torch-audio  # 音声処理が必要な場合のみ

# Development dependencies (optional)
# pytest>=6.0.0
# black>=21.0.0
# flake8>=3.8.0